Update rgpluaclassbrowser.lua
Add searching also by matching any part of the class name (method, etc.), and not just by the beginning.
search_text does not need to be checked for nil here because include_all precedes this one. It is desirable to move string.lower(search_text) outside the for loop, but then it is necessary to prevent a null argument for string.lower().